@@ -9,7 +9,7 @@
         }
     ],
     "require": {
-		"php": ">=5.6.0",
+        "php": "^7.4|^8.0",
     	"themeplate/core": "*"
     },
     "autoload": {
